Fix support for descendant selectors.
Fixes a regression in v1.3.0 where descendant selectors were possible
because all listener selectors where always checked against the document
using document.querySelectorAll() irrespective of the type of mutation
and whether the nodes were added or removed.

After the change to consult MutationRecord.addedNodes only the top-level
node of whole added subtrees are checked, and use of Element.matches()
restricts to those elements that match listener.selector and all
children in the subtree are now ignored. Again in the past this worked
because the selector was queried from the document.

The fix is to take the element if it matches, but also add all those
elements that are decendants by using element.querySelectorAll(). This
should solve most situations that were previously supported, but not
where the added node is not an ancestor of the selector provided.

In addition, the check for nodeType has been moved ahead of looping over
the listeners to avoid unnecessary cycles for other node types.